ducklake-import-export/pyproject.toml
[project]
name = "ducklake-import-export"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9",
]
